inserir e commit

+++ src/CursoOnline.Dominio/Cursos/CursoDto.cs
@@ -5,9 +5,9 @@
     {
         public string Nome { get; set; }
         public string Descricao { get; set; }
-        public decimal CargaHoraria { get; set; }
+        public double CargaHoraria { get; set; }
         public string PublicoAlvo { get; set; }
-        public decimal Valor { get; set; }
+        public double Valor { get; set; }
         public int Id { get; set; }
     }
 }
